WeeklyEventStorage: Remove templates.
Specialize WeeklyEventStorage to use the basic `int` type for
event codes, rather than being a template class. The removes
type information at call sites using an `enum class` but also
reduces code bloat, which we prefer.
